user role

+++ modules/Bido/Category/Http/Controllers/CategoryController.php
@@ -2,10 +2,9 @@
 
 namespace Bido\Category\Http\Controllers;
 
-use Illuminate\Http\Response;
 use Bido\Category\Models\Category;
 use App\Http\Controllers\Controller;
-use Bido\Category\Responses\AjaxResponses;
+use Bido\Common\Responses\AjaxResponses;
 use Bido\Category\Repositories\CategoryRepo;
 use Bido\Category\Http\Requests\CourseRequest;
 use Bido\Category\Http\Requests\CategoryRequest;
@@ -21,6 +20,7 @@ class CategoryController extends Controller
 
     public function index()
     {
+        $this->authorize('manage', Category::class);
         $categories = $this->repo->all();
 
         return view('Categories::index', compact('categories'));
@@ -28,12 +28,14 @@ class CategoryController extends Controller
 
     public function store(CategoryRequest $request)
     {
+        $this->authorize('manage', Category::class);
         $this->repo->store($request);
         return back();
     }
 
     public function edit($categoryId)
     {
+        $this->authorize('manage', Category::class);
         $category = $this->repo->findById($categoryId);
 
         $categories = $this->repo->allExceptById($categoryId);
@@ -43,6 +45,7 @@ class CategoryController extends Controller
 
     public function update(CategoryRequest $request, $categoryId)
     {
+        $this->authorize('manage', Category::class);
         $this->repo->update($categoryId, $request);
 
         return back();
@@ -50,6 +53,7 @@ class CategoryController extends Controller
 
     public function destroy($categoryId)
     {
+        $this->authorize('manage', Category::class);
         $this->repo->delete($categoryId);
 
         return AjaxResponses::successResponse();
