+++ modules/Bido/RolePermissions/Http/Controllers/RolePermissionsController.php
@@ -2,14 +2,15 @@
 
 namespace Bido\RolePermissions\Http\Controllers;
 
-use Spatie\Permission\Models\Permission;
-use Bido\Category\Responses\AjaxResponses;
+use App\Http\Controllers\Controller;
+use Bido\RolePermissions\Models\Role;
+use Bido\Common\Responses\AjaxResponses;
 use Bido\RolePermissions\Repositories\RoleRepo;
 use Bido\RolePermissions\Http\Requests\RoleRequest;
 use Bido\RolePermissions\Repositories\PermissionRepo;
 use Bido\RolePermissions\Http\Requests\RoleUpdateRequest;
 
-class RolePermissionsController
+class RolePermissionsController extends Controller
 {
     private $roleRepo;
     private $permissionRepo;
@@ -22,6 +23,7 @@ class RolePermissionsController
 
     public function index()
     {
+        $this->authorize('index', Role::class);
         $roles = $this->roleRepo->all();
         $permissions = $this->permissionRepo->all();
 
@@ -30,11 +32,15 @@ class RolePermissionsController
 
     public function store(RoleRequest $request)
     {
-       return $this->roleRepo->create($request);
+        $this->authorize('create', Role::class);
+        $this->roleRepo->create($request);
+
+        return redirect(route('role-permissions.index'));
     }
 
     public function edit($roleId)
     {
+        $this->authorize('edit', Role::class);
         $role = $this->roleRepo->findById($roleId);
         $permissions = $this->permissionRepo->all();
 
@@ -43,6 +49,7 @@ class RolePermissionsController
 
     public function update(RoleUpdateRequest $request, $id)
     {
+        $this->authorize('edit', Role::class);
        $this->roleRepo->update($request, $id);
 
        return redirect(route('role-permissions.index'));
@@ -51,6 +58,7 @@ class RolePermissionsController
 
     public function destroy($roleId)
     {
+        $this->authorize('delete', Role::class);
         $this->roleRepo->delete($roleId);
         return AjaxResponses::successResponse();
     }
